add script notebook and the resul of the run
