--- successfully-unsuccessful-c8dd9ca7a392.txt
@@ -1,112 +1,0 @@
-url: https://medium.com/@sushant.waghmode/successfully-unsuccessful-c8dd9ca7a392
-
-Dr Sushant
-
-Follow
-
---
-
-Listen
-
-Share
-
-Successfully unsuccessful is what they call me.
-
-And I guess I finally get it why. It’s not that I haven’t emasculated my own consciousness by engaging myself into each and every particulate that came along. The amount of energy I have invested in trying to learn things only out of the expectation that one of them would be my golden ticket out of this mediocre life span. I have been consistent, disciplined, passionate and patient. I have tried to cultivate every attribute relevant or irrelevant so that one day all the pieces would come together to make a beautiful masterpiece.
-
-But I was successfully unsuccessful.
-
-Does that mean I give up? Is this the end? Have I really run out of rope to hold on to? Is this the demise of a soul that tried but never could be? Have the clutches of destiny finally strangled all hope that was left. In the end this material world has prevailed and concurred upon the actions of my soul. Even after all that effort the bottomless pit of failure is where I find myself. Falling endlessly into the void. Torn apart into fragments by the gravitational pull of my failures. But I tried so hard is this really the end. Is this how this story ends?
-
-Nay I say. Never!!
-
-Until there is a fragment. Until there is a molecule, an atom of my existence, I will return. Each fragment will spontaneously combust like a self-illuminating star and purge upon the darkness. This holy war upon the inevitable will last for eternities to come. For the truth is that I will never give up. Rise from the ashes I will. The winds will howl by my name and I will stare into that bottomless pit and shout into the abyss. You shall not have me! Not now, not today, not tomorrow, not in any time frame. I will fail an infinite times until I chip at your fortress block by block. And in that moment when you presume that I am done I will rise high like a supernova and make my mark upon this world.
-
-Successfully unsuccessful, but not yet done.
-
---
-
---
-
-Dentist, Councillor, Fitness enthusiast, Poet
-
-Dr Sushant
-
---
-
-1
-
-Dr Sushant
-
---
-
-1
-
---
-
-1
-
-AL Anany
-
---
-
-161
-
-Unbecoming
-
---
-
-931
-
-Deepa Gaitonde
-
---
-
-2
-
-Nick Hilton
-
---
-
-133
-
-Victor Timi
-
-in
-
-Level Up Coding
-
---
-
-17
-
-Zulie Rane
-
-in
-
-The Startup
-
---
-
-916
-
-Help
-
-Status
-
-Writers
-
-Blog
-
-Careers
-
-Privacy
-
-Terms
-
-About
-
-Text to speech
-
-Teams
-
